+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 ## 0.3.7
+
 * fix: `props.stripe` 在虚拟滚动是否开启时表现不同的问题。
 * fix: td `line-height` 为 1 导致英文基线下方部分被截断。现设置为默认值。
+* fix: `highlightDim` 方法，修复高亮不返回当前背景颜色的问题。
+
 ## 0.3.6
 
 * fix: 列数从多变少时，virtualScrollX.endIndex 超出数组长度的问题。